+++ packages/gamut/src/RadialProgress/index.tsx
@@ -18,6 +18,10 @@ const offsetDelta = offsetForEmptyProgress - offsetForFullProgress;
 const convertPercentToOffset = (percent: number) =>
   offsetForEmptyProgress - Math.floor(offsetDelta * (percent / 100));
 
+/**
+ * @deprecated
+ */
+
 export const RadialProgress: React.FC<RadialProgressProps> = ({
   children,
   className,