+++ packages/styleguide/.storybook/preview.ts
@@ -1,11 +1,11 @@
 import 'focus-visible/dist/focus-visible.min.js';
 
-import './decorators/wrapper';
 import { withEmotion } from './decorators/theme';
-import { Page } from './components';
-import { colors } from '@codecademy/gamut-styles/src';
+import { DocsPage, DocsContainer } from './components';
+import { theme as gamutTheme } from '@codecademy/gamut-styles/src';
 import { theme } from './theme';
-import { breakpoints } from '@codecademy/gamut-styles';
+
+const { colors, breakpoints } = gamutTheme;
 
 export const parameters = {
   viewMode: 'docs',
@@ -18,7 +18,9 @@ export const parameters = {
         'Foundations',
         ['About', 'Theme', 'System', 'Design Guidelines', 'Legacy'],
         'Typography',
+        ['About', 'Text', 'Anchor'],
         'Layouts',
+        ['About', 'Box', 'LayoutGrid', 'ContentContainer'],
         'Atoms',
         'Molecules',
         'Organisms',
@@ -31,16 +33,21 @@ export const parameters = {
       locales: 'en-US',
     },
   },
+  taxonomy: {
+    root: 'gamut',
+    indexPage: 'about',
+  },
   // Addon Options
   docs: {
     theme,
+    container: DocsContainer,
     components: {
-      wrapper: Page,
+      wrapper: DocsPage,
     },
   },
   backgrounds: {
     grid: {
-      cellSize: 20,
+      cellSize: 16,
       opacity: 0.5,
       cellAmount: 5,
     },
@@ -53,6 +60,10 @@ export const parameters = {
   viewport: {
     defaultViewport: 'responsive',
     viewports: {
+      responsive: {
+        name: 'Responsive',
+        type: 'desktop',
+      },
       xs: {
         name: `XS - ${breakpoints.xs}`,
         styles: {