+++ packages/styleguide/CHANGELOG.md
@@ -3,6 +3,227 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [33.7.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.7.0...@codecademy/styleguide@33.7.1) (2021-04-06)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.7.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.6.3...@codecademy/styleguide@33.7.0) (2021-04-06)
+
+
+### Features
+
+* **Docs:** Docs refresh (Grid, Legacy Colors, Typography) [GM-118] [GM-194]  ([623e7b3](https://github.com/Codecademy/client-modules/commit/623e7b3c4fd3b22006af5dc19780aadb2f235f18))
+
+
+
+### [33.6.3](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.6.2...@codecademy/styleguide@33.6.3) (2021-04-05)
+
+
+### Bug Fixes
+
+* Typo... ([16a6eec](https://github.com/Codecademy/client-modules/commit/16a6eecf9e8bedefcec73d26f84af2bc456c80a2))
+
+
+
+### [33.6.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.6.1...@codecademy/styleguide@33.6.2) (2021-04-05)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.6.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.6.0...@codecademy/styleguide@33.6.1) (2021-04-05)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.6.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.5.0...@codecademy/styleguide@33.6.0) (2021-04-01)
+
+
+### Features
+
+* **Input:** React-Recurly edits ([1d2a286](https://github.com/Codecademy/client-modules/commit/1d2a28670b6b218a891462191a80e97a982482c9))
+
+
+
+## [33.5.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.4.2...@codecademy/styleguide@33.5.0) (2021-04-01)
+
+
+### Features
+
+* **GridForm:** Update the buttons for GridForm submit ([cb03dd2](https://github.com/Codecademy/client-modules/commit/cb03dd2a7bce0d2b9baa99349bae22824c0e614b))
+
+
+
+### [33.4.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.4.1...@codecademy/styleguide@33.4.2) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.4.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.4.0...@codecademy/styleguide@33.4.1) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.4.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.9...@codecademy/styleguide@33.4.0) (2021-03-30)
+
+
+### Features
+
+* **Text:** Deprecate Text / Heading and add more robust UI Text Component [GM-162] ([78e284f](https://github.com/Codecademy/client-modules/commit/78e284f3ce80f14e583af3133f692a7cd91e2196))
+
+
+
+### [33.3.9](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.8...@codecademy/styleguide@33.3.9) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.8](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.7...@codecademy/styleguide@33.3.8) (2021-03-29)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.7](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.6...@codecademy/styleguide@33.3.7) (2021-03-26)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.6](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.5...@codecademy/styleguide@33.3.6) (2021-03-26)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.5](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.4...@codecademy/styleguide@33.3.5) (2021-03-25)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.4](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.3...@codecademy/styleguide@33.3.4) (2021-03-25)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.3](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.2...@codecademy/styleguide@33.3.3) (2021-03-25)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.1...@codecademy/styleguide@33.3.2) (2021-03-24)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.3.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.3.0...@codecademy/styleguide@33.3.1) (2021-03-24)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.2.1...@codecademy/styleguide@33.3.0) (2021-03-23)
+
+
+### Features
+
+* **Input:** Add custom gamut-icons to Input components ([cfb3fbe](https://github.com/Codecademy/client-modules/commit/cfb3fbe390a88f7e2e513bc76ef90a54051de665))
+
+
+
+### [33.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.2.0...@codecademy/styleguide@33.2.1) (2021-03-23)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.1.0...@codecademy/styleguide@33.2.0) (2021-03-23)
+
+
+### Features
+
+* **GamutProvider:** Adds single provider for Theme + Cache + Globals + Assets [WEB-1352] [GM-173] ([9eac700](https://github.com/Codecademy/client-modules/commit/9eac7009b4d773621a0993cc125d6878b76977ce))
+
+
+
+## [33.1.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.0.2...@codecademy/styleguide@33.1.0) (2021-03-23)
+
+
+### Features
+
+* **Dialog:** Add new Dialog Modal component [GM-18] ([349faa9](https://github.com/Codecademy/client-modules/commit/349faa911f61fbf08aa14234d8dfc7ae50135fd7))
+
+
+
+### [33.0.2](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.0.1...@codecademy/styleguide@33.0.2) (2021-03-22)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+### [33.0.1](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@33.0.0...@codecademy/styleguide@33.0.1) (2021-03-22)
+
+**Note:** Version bump only for package @codecademy/styleguide
+
+
+
+
+
+## [33.0.0](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@32.0.3...@codecademy/styleguide@33.0.0) (2021-03-22)
+
+
+### ⚠ BREAKING CHANGES
+
+* converted ToolTip to Emotion (#1409)
+
+### Code Refactoring
+
+* converted ToolTip to Emotion ([#1409](https://github.com/Codecademy/client-modules/issues/1409)) ([0a8428f](https://github.com/Codecademy/client-modules/commit/0a8428fb59998e98d053b880c92d884fe5639a1d))
+
+
+
 ### [32.0.3](https://github.com/Codecademy/client-modules/compare/@codecademy/styleguide@32.0.2...@codecademy/styleguide@32.0.3) (2021-03-19)
 
 **Note:** Version bump only for package @codecademy/styleguide