+++ packages/variance/CHANGELOG.md
@@ -3,6 +3,67 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.5.0...@codecademy/variance@0.6.0) (2021-04-06)
+
+
+### Features
+
+* **Variance:** Initial Props and Refactor Tests [WEB-1390] ([e4b5177](https://github.com/Codecademy/client-modules/commit/e4b5177544643fec274d20f37be594b140a4cba3))
+
+
+
+## [0.5.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.4.0...@codecademy/variance@0.5.0) (2021-03-31)
+
+
+### Features
+
+* **Variance:** Use only the global emotion theme type (Stage 5: Acceptance) [WEB-1382] ([e4215e9](https://github.com/Codecademy/client-modules/commit/e4215e9aacb2092ed77cd530822c5389c7b47123)), closes [/github.com/Codecademy/client-modules/blob/609bc7f3ab936d856225fb9885d0ae5afeac8b58/packages/variance/src/__tests__/localTheme.ts#L28-L29](https://github.com/Codecademy//github.com/Codecademy/client-modules/blob/609bc7f3ab936d856225fb9885d0ae5afeac8b58/packages/variance/src/__tests__/localTheme.ts/issues/L28-L29)
+
+
+
+## [0.4.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.3.0...@codecademy/variance@0.4.0) (2021-03-31)
+
+
+### Features
+
+* **Variance:** base => _ for un scoped style [WEB-1381] ([f3d254c](https://github.com/Codecademy/client-modules/commit/f3d254cccc3dc3cd074df070a51ed8a2ca232a74))
+
+
+
+## [0.3.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.2.1...@codecademy/variance@0.3.0) (2021-03-31)
+
+
+### Features
+
+* **Variance:** `css` - Standard CSS Property Passthrough + Selector System Props [WEB-1117] ([6c2c48d](https://github.com/Codecademy/client-modules/commit/6c2c48dc96012bc685ca4ab93d189d1e39d266b1))
+
+
+
+### [0.2.1](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.2.0...@codecademy/variance@0.2.1) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/variance
+
+
+
+
+
+## [0.2.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.1.2...@codecademy/variance@0.2.0) (2021-03-30)
+
+
+### Features
+
+* **Variance:** Static CSS Methods `css` and `variant`  [WEB-1116] ([b9915f5](https://github.com/Codecademy/client-modules/commit/b9915f54bc738d81e46cb84744a44e8bfbdd29a0))
+
+
+
+### [0.1.2](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.1.1...@codecademy/variance@0.1.2) (2021-03-30)
+
+**Note:** Version bump only for package @codecademy/variance
+
+
+
+
+
 ### [0.1.1](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.1.0...@codecademy/variance@0.1.1) (2021-03-19)
 
 **Note:** Version bump only for package @codecademy/variance